dns: add a wildcard record for blackfeminism
this is to test alias records to an existing ELB. the name and zone id provided
target the kubernetes cluster's elastic load balancer, introducing this record
should allow us to test that `something.blackfeminism.library.ucsb.edu` reaches
the cluster's ingress controller.

if this works, we can change the record for the top level to complete the DNS
changeover.