move AWS load balancer targets to the new DLD ingress controller
DLD had been running an older version of
https://kubernetes.github.io/ingress-nginx/, which supported both the v1beta
Ingress API and the newer v1 API.

With the upgrade to Kubernetes v1.22, the v1beta API has been removed, which is
causing that ingress controller deployment to fail when assigning new
ingresses (it's trying to list all v1beta ingresses, but the resource definition
doesn't exist).

We have a newer Ingress Controller in place which is working, running version
1.x of the controller application. The new load balancer targeted here is the
one that points to the new ingress controller.