Mv alexandria to lb haproxy legacy 001 (#65)
* Update zone.alexandria.ucsb.edu.tf

Changing the CNAME record for www.alexandria.ucsb.edu to point to lb-haproxy-legacy-001.library.ucsb.edu.  Changing the A record for alexandria.ucsb.edu to a CNAME to point to lb-haproxy-legacy-001.library.ucsb.edu.

* Corrected a syntax error

Needed to add the trailing "." in lb-haproxy-legacy-001.library.ucsb.edu. for the alexandria.ucsb.edu CNAME.  Strangely enough, the Initialize and Plan did not deem this to result in an error.